Changed name to autoware_bag_tools and added README
