+++ ros/src/util/packages/autoware_bag_tools/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(bag_tools)
+project(autoware_bag_tools)
 
 find_package(catkin REQUIRED COMPONENTS rospy rosbag)
