+++ ros/src/util/packages/autoware_bag_tools/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>bag_tools</name>
+  <name>autoware_bag_tools</name>
   <version>1.0.0</version>
   <description>Autoware modified bag_tools; ROS tools and scripts related to bagfiles</description>
   <maintainer email="jacob.lambert@tier4.jp">Jacob Lambert</maintainer>
